src/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
SQLAlchemy==2.0.38
Flask-Migrate==4.0.5
python-dotenv==1.0.1
selenium==4.32.0
